Asynchronous Operations using Tasks (async, await) Corrected

+++ SalesWebMvc/Services/SellerService.cs
@@ -28,15 +28,21 @@ namespace SalesWebMvc.Services
             await _context.SaveChangesAsync();
         }
 
-        public async Task<Seller> FindByIdAsync(int id)
+        public async Task<Seller?> FindByIdAsync(int id)
         {
-            return await _context.Seller.Include(obj => obj.Department).FirstOrDefaultAsync(obj => obj.Id == id);
+            return await _context.Seller
+                .Include(obj => obj.Department)
+                .FirstOrDefaultAsync(obj => obj.Id == id);
         }
 
         public async Task RemoveAsync(int id)
         {
             var obj = await _context.Seller.FindAsync(id);
-            
+            if (obj == null)
+            {
+                throw new NotFoundException("Id not found");
+            }
+
                 _context.Seller.Remove(obj);
                 await _context.SaveChangesAsync();
         }
